requirements.txt
altair==5.5.0
ansicon==1.89.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blessed==1.20.0
blinker==1.9.0
Brotli==1.1.0
browser-cookie3==0.19.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.2
commonmark==0.9.1
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
fastjsonschema==2.21.1
Flask==3.1.0
fonttools==4.53.1
fqdn==1.5.1
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
Health==1.4
httpcore==1.0.7
httpx==0.28.0
idna==3.7
ipykernel==6.29.5
ipython==8.30.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jinxed==1.3.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
lxml==4.9.4
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
measurement==3.2.2
mistune==3.0.2
mpmath==1.3.0
mutagen==1.47.0
myfitnesspal==2.1.0
narwhals==1.15.2
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.1
notebook_shim==0.2.4
numpy==1.26.4
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.1
pafy==0.5.5
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pycryptodomex==3.20.0
pydeck==0.9.1
Pygments==2.18.0
pyjokes==0.6.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-vlc==3.0.20123
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rpds-py==0.22.3
scikit-learn==1.5.2
scipy==1.13.1
Send2Trash==1.8.3
setuptools==75.6.0
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.40.2
sympy==1.13.3
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.0
yt-dlp==2024.7.25
matplotlib==3.9.1
pandas==1.5.3
streamlit==1.25.0
Pillow==10.0.0
requests==2.31.0
